Adding a new filter called “pmpro_stripe_create_subscription_array” inside class.pmprogateway_stripe.php, LINE 1409.
This is to allow other plugins to modify the subscription object before sending to Stripe. This is most useful for setting a custom ‘trial_end’ timestamp inside Stripe (which cannot be done inside the Stripe plan object).

https://stripe.com/docs/api/php#create_subscription

The Stripe plan object only lets you set a number of ‘trial_period_days’ — which is not granular enough if you want subscriptions to renew at a specific time of day — like noon or 3pm or whatever.