Start and enddate bugfix
My client told me that the start and end date are not working on the cancel_admin email. So I figured out that the current code only looks for dates from users with status = inactive, but we had a lot of people who had status cancceled or admin_cancelled and therefore the cancel_admin email did not find the right post.

I do not know if this is the right fix for this but it worked for us!