Add Stripe class mods to handle MMPU.
The token needs to be different for each run through, so the javascript
now generates a new one per level, and attaches it to the form as
stripeToken0, stripeToken1, etc. Then, on processing, the class goes
through $_REQUEST to find the highest-numbered one, uses it, and drops
it off of $_REQUEST. That way, each subscription/payment has a novel
token.