BUG/FIX: PayPal Standard/Add-on Packages issue
The add-on packages expects to handle assigning a page to the user via the checkout preheader hooks, the PayPal Standard gateway returns the user to the confirmation page. Also, the current way of handling the PayPal arguments for the request assumes that the return URL is the last entry in the NVP string. That's not correct.