Moved template filter closer to sendEmail call
Currently, the template filter is never invoked for free level
checkout, which seems like an incomplete implementation, and doesn’t
give any opportunity to override the template for free-level or trial
checkout. Moving the filter invocation lower makes it accessible to all
(but still keeps it before the call to sendEmail).