Enhancement: Improved level setting warnings
Enhancement: Warn users when a level has a recurring subscription and an expiration date. Highlight the level in red when this occurs inside the level settings page and show a error message inside the edit level page.

Enhancement: Bolden the text under Recurring Subscription to make it more noticeable.

Enhancement: Filter added to adjust the pmpro_checkLevelForSettingsCompatibility function. This function is used