Show if plugin is active.
Rule to show notice if plugin is active.

I think we should abstract this to return true or false if the plugin is active. That way we can recall this multiple times.

Also before checking each step, if it's already set to true, we should just show the message instead of looping vs everything.

The logic behind here is:

* Check the rules to hide the notice.
* Check the rules to show the notice, if the first one is true show the notice.