Update class.memberorder.php
In saveOrder when we check if the tax has be set, we check if isset() and if the tax is defaulted to 0, that's going to return true, which would lead to us calculating taxes every time the order is saved, which could be costly if we're hitting an API to calculate taxes. So I set the tax property to default to null.