Idea: Filter to hide the setup wizard / disable redirect.
Added in a filter to show/hide the setup wizard based on custom conditions. I assume web developers won't want their customer's running through the setup wizard in some instances and want to hide it from them as things may be overwritten or added when running through the setup wizard?