Fixed always redirect to login page
* BUG FIX: Fixed an issue where the redirect would always redirect to the login page.

The issue here was we were ommitting the redirect value entirely from the $login_form_array which would then default to setting a $redirect_to value of the REQUEST URI and it would always redirect to that URL based on the wp_login_form defaults when that parameter was missing.

This fix, fixes issues for custom redirects (snippets) as well as the Member Homepages Add On.