Comment out update_subscription
This is a temporary fix to remove the update_subscription method from the base gateway class as we have checks to see if the method_exists for a gateway object.

This will always return true since we are extending the base class and `method_exists` isn't smart enough to know if we're overriding the method or not within our actual gateway used class.

Gateways may still declare the `cancel_subscription` method and use this as the preffered/newer way of doing things and gateways that are still using the deprecated `cancel()` method would still work.